Renamed physicalvalidation to physical_validation, partially fixes issue #5.
