+++ README.md
@@ -1,13 +1,13 @@
 Physical validation reference
 =============================
 
-`physicalvalidation` is a package aimed at testing results obtained
+`physical_validation` is a package aimed at testing results obtained
 by molecular dynamics simulations for their physical validity.
 
 Please check [https://shirtsgroup.github.io/physical-validation](https://shirtsgroup.github.io/physical-validation)
 for the full reference.
 
-`physicalvalidation` largely incorporates the functionality of 
+`physical_validation` largely incorporates the functionality of
 [checkensemble](https://github.com/shirtsgroup/checkensemble).
 
 This software is developed in the Shirts group at University of
