+++ doc/physical_validation.data.rst
@@ -1,0 +1,26 @@
+physical_validation\.data subpackage
+====================================
+
+physical_validation\.data\.simulation_data module
+-------------------------------------------------
+.. automodule:: physical_validation.data.simulation_data
+    :members:
+    :undoc-members:
+
+
+physical_validation\.data\.parser module
+----------------------------------------
+.. automodule:: physical_validation.data.parser
+
+.. autoclass:: physical_validation.data.parser.Parser
+    :members:
+    :undoc-members:
+
+
+physical_validation\.data\.gromacs_parser module
+------------------------------------------------
+.. automodule:: physical_validation.data.gromacs_parser
+
+.. autoclass:: physical_validation.data.gromacs_parser.GromacsParser
+    :members:
+    :undoc-members:
