+++ doc/physical_validation.rst
@@ -1,0 +1,23 @@
+physical_validation package
+===========================
+
+.. automodule:: physical_validation
+    :members:
+
+physical_validation\.kineticenergy module
+-----------------------------------------
+
+.. automodule:: physical_validation.kineticenergy
+    :members:
+
+physical_validation\.ensemble module
+------------------------------------
+
+.. automodule:: physical_validation.ensemble
+    :members:
+
+physical_validation\.integratorconvergence module
+-------------------------------------------------
+
+.. automodule:: physical_validation.integrator
+    :members:
