--- doc/physicalvalidation.data.rst
@@ -1,26 +1,0 @@
-physicalvalidation\.data subpackage
-===================================
-
-physicalvalidation\.data\.simulation_data module
-------------------------------------------------
-.. automodule:: physicalvalidation.data.simulation_data
-    :members:
-    :undoc-members:
-
-
-physicalvalidation\.data\.parser module
----------------------------------------
-.. automodule:: physicalvalidation.data.parser
-
-.. autoclass:: physicalvalidation.data.parser.Parser
-    :members:
-    :undoc-members:
-
-
-physicalvalidation\.data\.gromacs_parser module
------------------------------------------------
-.. automodule:: physicalvalidation.data.gromacs_parser
-
-.. autoclass:: physicalvalidation.data.gromacs_parser.GromacsParser
-    :members:
-    :undoc-members:
