--- doc/physicalvalidation.rst
@@ -1,23 +1,0 @@
-physicalvalidation package
-==========================
-
-.. automodule:: physicalvalidation
-    :members:
-
-physicalvalidation\.kineticenergy module
-----------------------------------------
-
-.. automodule:: physicalvalidation.kineticenergy
-    :members:
-
-physicalvalidation\.ensemble module
------------------------------------
-
-.. automodule:: physicalvalidation.ensemble
-    :members:
-
-physicalvalidation\.integratorconvergence module
-------------------------------------------------
-
-.. automodule:: physicalvalidation.integrator
-    :members:
