--- doc/physicalvalidation.util.rst
@@ -1,42 +1,0 @@
-physicalvalidation\.util subpackage
-===================================
-
-.. warning:: This subpackage is intended for internal use. Documentation
-    and input validation is reduced.
-
-
-physicalvalidation\.util\.checkensemble module
-----------------------------------------------
-.. automodule:: physicalvalidation.util.checkensemble
-    :members:
-    :undoc-members:
-
-physicalvalidation\.util\.error module
---------------------------------------
-.. automodule:: physicalvalidation.util.error
-    :members:
-    :undoc-members:
-
-physicalvalidation\.util\.gromacs_interface module
---------------------------------------------------
-.. automodule:: physicalvalidation.util.gromacs_interface
-    :members:
-    :undoc-members:
-
-physicalvalidation\.util\.integrator module
--------------------------------------------
-.. automodule:: physicalvalidation.util.integrator
-    :members:
-    :undoc-members:
-
-physicalvalidation\.util\.kineticenergy module
-----------------------------------------------
-.. automodule:: physicalvalidation.util.kineticenergy
-    :members:
-    :undoc-members:
-
-physicalvalidation\.util\.timeseries module
--------------------------------------------
-.. automodule:: physicalvalidation.util.timeseries
-    :members:
-    :undoc-members:
